requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
openai==1.40.0
httpx==0.27.2
pandas==2.2.2
openpyxl==3.1.5
numpy==1.26.4
werkzeug==3.0.4
